% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_history_trait}
\alias{life_history_trait}
\title{Life history trait}
\format{A data.frame with two variables: \code{species} and
\code{life_history_trait}}
\usage{
life_history_trait
}
\description{
Life history trait coded as eXotic/Native, Annual/Perennial, and D/M
}
\examples{
library(dplyr)
with_cover <- cover \%>\%
  left_join(life_history_trait, by="species")
}
\seealso{
\code{\link{cover}}
}
\keyword{datasets}

